Implement proper patch access control
We shouldn't be allowing anyone who happens to know the URL access to
the patch files uploaded as part of an image comparison, so use a proper
Django view coupled with nginx's X-Accel-Redirect to limit access to the
files to logged-in users.

In addition, mount the patchvolume as read-only in layersweb and use the
z flag on the layersapp mount of it; the ro flag in particular seemed to
fix a weird permission denied issue I was having just listing the
directories containing patch files.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>