oe-image-manifest: fix handling of remote patches
Some recipes (e.g. current bash 4.4.x recipe in OE-Core) pull in patch
files remotely. These are treated as remote sources, but were also being
processed as patches, except that processing assumed the files would be
located next to the recipe, and as a traceback would be the result if
that turned out to be false.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>