Add layer export script
Add a script that will export patches from packages in a comparison
branch to a layer, using bbappends to apply the patches when building
the corresponding recipe. Patches to export are filtered by being
associated with a package that is linked to a recipe on the OE side and
where the patch dispositioning does not exclude it. During the export,
patches in the comparison branch and existing patches in the recipe are
hashed and if the hashes match then we can exclude the patch, setting
the dispositioning field automatically. This is only sometimes
effective - we do chop off the header to ignore differences in that
part, but itdifferences in line numbers alone (which can of course be
fuzzed out when applying) will still be seen as a different patch.
Note that the layer to export to must already exist as a layer.

There is some logic in the script to print a note when the main source
URL in SRC_URI is not the same as the one in the comparison branch
package, but that doesn't do anything except report it.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>