Support version comparisons to images
Add support for comparing versions between image comparison branches or
a distro comparison and image comparison. This isn't perfect - the
distro comparison branch needs to be limited to the packages in an image
(which will be the case for the new derivative import functionality)
otherwise the version comparison will report lots of packages as having
been removed.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>