Fix diff status return when using nginx
When nginx is acting as the proxy, we use X-Accel-Redirect to get nginx to
return the content of the file (diff in this case). However, any headers
set on the original response don't get automatically transferred through
to the response that nginx sends to the client though, so in order to
still have our X-Status header that the client-side js looks for we need
to explicitly add it through using an add_header directive.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>