import_clear: fix derivative import restore of original data
When we do a derivative import we need to move the original distro data
source directory out of the way if it exists since we want to limit the
source packages to the ones in the derivative. When we move it back at
the end however we need to first delete the derivative source dir
otherwise we'll be moving the original sources underneath it.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>