import_clear: preserve derivative source
If we want to be able to do file comparisons we need to keep the package
source. This isn't difficult, we just need to move it to an appropriate
place instead of deleting it; we already store the path to it in the
local_path field so the file comparison code will know where to pick it
up.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>