docker: add dissector command-line tool to app container
The dissector command-line tool is needed for importing data from Clear
Linux; previously this needed to be manually built and mounted as a
volume into the layerscelery container. To avoid that, build and install
it when creating the app container and then throw away the Go tools
(which we don't need once the tool has been compiled).

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>