Handle release comparison of two image comparisons
Allow comparing two images imported via the image comparison feature
using the release comparison feature. To accomplish this we just needed
to link recipes directly by pn (instead of going via cover_pn) and then
for diff purposes get the code to look in the appropriate place to find
the patches. One caveat though is that we do not store the recipes when
exporting the manifest info so those won't be there to compare (unlike
spec files with imported Clear Linux data).

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>