export_layer: support fetching layer to export into
We need a layer to export into, and to make it easier in a web context,
allow specifying a git URL to fetch the layer from. This also requires a
subdirectory parameter since the layer might not be at the root of the
repo.

(Now that we're definitely importing into an existing layer, ensure that
we delete its contents prior to importing into it so that we only end up
with the newly generated content.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>