Add ability to associate files with an update
If an update process can generate files then we need a mechanism to
allow the user to download them. Add a variable UPDATE_FILES_DIR to
specify where these files should be stored, an UpdateFile model to
store a link to them, and the means to show the links to them in the
update detail and task pages (when the task is complete). The UpdateFile
objects can be created in advance and they will only show up if the file
actually exists at the end of the update.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>