Support updating recipes during export
Allow specifying that when exporting some packages, update an existing
recipe in the output layer instead of creating a bbappend. What this
will do:

* Set SRC_URI to point to the sources and patches specified in the
  package
* Copy the patches in
* For kernel recipes:
 - Update LINUX_VERSION if appropriate (for kernel recipes)
 - Update the defconfig

Note that we do not use the covering recipe in this case - we don't
even care whether it's set in this mode, we instead look for a recipe
with PN directly matching the package name.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>