Support saved searches for comparison patch search
Add a mechanism for saving searches in the patch search form which can
be recalled simply by clicking on a link. This is useful because
for example searching for all outstanding patches can be complicated and
even usage-specific (e.g. if a layer filtering needs to be applied).

(There's no means of creating these apart from through the admin
interface at the moment, but it's a start.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>