Add missing permissions checks to image comparison views
Some of these views weren't checking if the user had permission to view
(at the moment, that means they're either a superuser or they created
the comparison), so if you accessed a URL directly then you were able to
see the page and in one case you'd have been able to change the
association of an image comparison recipe. Add the missing checks to fix
this.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>