settings: fix reregistration page redirecting to login
The re-registration notification page is supposed to show when you
edit your profile, change your email address and then save, but when the
page is shown the current user is logged out but we did not have the URL
for that page in the whitelist, so it redirected you to the login page.
Add it to the whitelist to avoid the issue.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>