Validate json data
Fields with choices are not actually validated when you save an object -
normally this is handled within the form. When we're writing objects
based on user input we need to do this validation ourselves, so add it
in a few places where we are doing that. Also add validation for a few
other specific fields.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>