API: fix recipes API performance regression
Adding these extra child items to the standard "recipes" viewset (which
we did recently in 874601c14853c8b27aff37952be3b472fe515f23) means that
some current usage is impractical due to the size of the returned list
of items. Instead, create a recipesExtended viewset, move the new child
items to that and add pagination to avoid result size issues.

[Note: the API remains disabled in the Clear Linux Dissector, this is
just to keep the code in sync with the OE layer index].

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>