dockersetup: ensure letsencrypt volume change gets undone
If you enable the --letsencrypt option when you run dockersetup.py, the
script will modify the volume mount for the certificates to point to
/etc/letsencrypt instead of /opt/cert. If you then run dockersetup.py
again (with -r/--reinstall) without --letsencrypt, we want the path to
be set back to /opt/cert, so ensure that it does. Additionally, the code
wasn't actually setting the path for the layerscertbot service since
editing that section was done separately. (Admittedly, the letsencrypt
functionality has not been well-tested.)

Fixes https://github.com/intel/clear-linux-dissector-web/issues/12

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>