Decrease field length to avoid initial setup issue with MariaDB
I'm unsure as to why this wasn't a problem earlier, but attempting to
set up a new database from scratch in an Ubuntu container following the
steps in README.devel failed when at the "./manage.py migrate" step
with the following error:

  ...
  Applying layerindex.0028_branch_hidden... OK
  Applying layerindex.0029_imagecompare...Traceback (most recent call last):
  File "/opt/venv/lib/python3.6/site-packages/django/db/backends/utils.py", line 64, in execute
    return self.cursor.execute(sql, params)
  File "/opt/venv/lib/python3.6/site-packages/django/db/backends/mysql/base.py", line 101, in execute
    return self.cursor.execute(query, args)
  File "/opt/venv/lib/python3.6/site-packages/MySQLdb/cursors.py", line 206, in execute
    res = self._query(query)
  File "/opt/venv/lib/python3.6/site-packages/MySQLdb/cursors.py", line 312, in _query
    db.query(q)
  File "/opt/venv/lib/python3.6/site-packages/MySQLdb/connections.py", line 224, in query
    _mysql.connection.query(self, query)
MySQLdb._exceptions.OperationalError: (1071, 'Specified key was too long; max key length is 767 bytes')

Searching around seems to indicate that this is an explicit limit on the
size of the unique index; dropping the length from 255 down to 180 fixes
it (and we don't really care about the reduction in this case). We don't
need a new migration because existing databases are obviously OK or they
wouldn't have been able to be created, and it's the earlier migration
that triggers the problem in the first place so we can't avoid changing
that anyway.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>