Implement session timeouts
Problem:Inactive and unattended sessions can allow unauthorized
users to access resources without permission.
If a session expiration feature is not used,
accounts that have not been logged out of can be at risk of being hijacked,
even if the user is only away temporarily.

[Paul] fixed logic and added a message so that the user knows what
happened if the session has timed out.

Signed-off-by: Patnana, Venkata Sai <venkata.sai.patnana@intel.com>
Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>