tasks: adjust result codes to help track down failure
At the moment we are sometimes seeing the Clear Linux import process
finish with "terminated (-1)" on the frontend with no error message
in the output. To try to track down why this is happening, change the
result codes so that we can tell the difference between various error
states. High arbrary values (-95 and -96) have been used to avoid the
possibility of clashing with typical errno values.

(See https://github.com/intel/clear-linux-dissector-web/issues/17 for
further details.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>