tasks: squash CRs in output before saving to database
In a follow-up to b67c807ce9c8ae5dafc5ca04c9aaee409f7b9db4, when we are
saving the output from an update task to the database just after the
task finishes, due to the in-line updates using CRs (e.g. from download
progress), the log may be large enough to overrun the current 128MB
limit on packet size for requests to the database (max_allowed_packet
parameter in MariaDB), with the result that MariaDB terminates the
connection and we see the following in the celery logs:

django.db.utils.OperationalError: (2006, 'MySQL server has gone away')

This means that the task status did not get saved to the database and
the process return code was not sent back to the client (thus
"terminated (-96)" or earlier "terminated (-1)" in the web UI even if
the task succeeded). If we squash out all of the unnecessary progress
updates then the logs come down to a reasonable size and there's no
problem.

Fixes: https://github.com/intel/clear-linux-dissector-web/issues/17
Fixes: https://github.com/intel/clear-linux-dissector-web/issues/18

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>