Fix backup instructions when large logs are recorded
If the database contains task logs that are over the default max packet
size for MariaDB, then attempting to dump the database will fail, but it
won't be immediately obvious that that has happened - the gzipped sql
file will simply be truncated and have an error in it. To fix the
underlying issue, add an option to the example command for database
backup to increase the max packet size to 512MB. (Restoring the database
doesn't seem to suffer the same issue).

In future we should probably look at writing a script to do this and
in it properly check the return of mysqldump so that any problems get
noticed immediately.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>